+++ src/a_color.h
@@ -8,7 +8,6 @@ public:
   a_color (int);
   bool is_visible();
   bool is_transparent();
-  bool has_alpha();
   std::string color();
   std::string opacity();